refactor(point): Change usePoints parameter to Long
- 타입 통일을 위해 amount 파라미터를 Long으로 변경

+++ src/main/java/com/golfRental/domain/point/service/command/PointCommandService.java
@@ -7,5 +7,5 @@ public interface PointCommandService {
 
     PointBalanceResponse getBalance(Long userId);
 
-    PointUseResponse usePoints(Long userId, Integer amount);
+    PointUseResponse usePoints(Long userId, Long amount);
 }
